[feat] LottoGame 클래스 구현
LottoGame List<LottoNumbers>를 감싸는 1급콜렉션 역할을 수행하는 클래스 구현
LottoResult 수익 계산 메서드 추가
ResultCount 문서화 주석 추가
OutputView 출력 메서드 재설계. LottoResult를 받아 printResult 메서드 수행

+++ src/main/java/domain/lottoresult/ResultCount.java
@@ -13,7 +13,14 @@ public class ResultCount {
         count++;
     }
 
-    public long multiply(int operand) {
+    /**
+     * multiply는 ResultCount 인스턴스 값에 아떤 정수를 곱할 때 사용되는 메서드이다.
+     * 실제 프로젝트에서는, 수익금을 계산할 때 사용된다.
+     *
+     * @param operand 곱셈 연산을 수행할 정수이다.
+     * @return long으로 곱셈을 수행한 결과을 반환한다.
+     */
+    public long multiply(long operand) {
         return (long) count * operand;
     }
